+++ app/Providers/AppServiceProvider.php
@@ -8,11 +8,13 @@ use App\Models\AgencyType;
 use App\Models\Country;
 use App\Models\District;
 use App\Models\Division;
+use App\Models\Project;
 use App\Models\Upazila;
 use App\Models\User;
 use App\Models\Ward;
 use App\Policies\AgencyPolicy;
 use App\Policies\ManageReferenceDataPolicy;
+use App\Policies\ProjectPolicy;
 use App\Policies\UserPolicy;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
@@ -41,5 +43,6 @@ class AppServiceProvider extends ServiceProvider
         Gate::policy(Ward::class, ManageReferenceDataPolicy::class);
         Gate::policy(AgencyType::class, ManageReferenceDataPolicy::class);
         Gate::policy(Agency::class, AgencyPolicy::class);
+        Gate::policy(Project::class, ProjectPolicy::class);
     }
 }